Fix minor mismatch on a specific EF Core migration
Obs.:
If this were already in Prod, we should not touch it of course cause the EF Core migration would've been applied on the DB side by this time!

It's not the case, so we just revert/rollback the DB and reapply all migrations once again.
Commands used for that:
'Update-Database -Migration 0' - all migrations will be reverted in the database
'Update-Database' - will apply all EF migrations again, now with the proper name in place

LEARNING AND REGISTERING HERE
1) Reverting to an earlier migration and/or reverting the last one
Update-Database -Migration SpecificMigrationName
This command reverts the DB to this specified migration name, unapplying/reverting all the other EF Core migrations after that - CAREFUL!

2) Package Manager Console x CLI
- 'Update-Database' command is to be used with the Package Manager Console when you have the Microsoft.EntityFrameworkCore.Tools package installed in your prj
- 'dotnet ef database update migration SpecificMigrationName' command is to be used when you have the Microsoft.EntityFrameworkCore.Tools.DotNet